requirements.txt
# Base dependencies
numpy==1.24.3
moviepy==1.0.3
Pillow>=10.0.0
requests==2.31.0

# Text to Speech
gTTS==2.3.2

# AI/ML APIs
google-generativeai==0.3.1
replicate==0.22.0

# TikTok & Browser automation
TikTokApi>=5.2.2
playwright==1.40.0

# Utils
python-dotenv==1.1.0

# Tambahkan
nest-asyncio==1.5.8